Ranged Weapon Class, Refactoring
Added ranged weapon child class based on weapon.
Refactored weapon class with more generic terms.
Moved ranged based weapon logic into ranged weapon script.
