+++ Assets/Scripts/Combat/RangedWeapon.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: a77b92ae1939e1d48ac6ddd08347e4d8
